Rename classes to make naming more coherent
