+++ rsmf/__init__.py
@@ -1,1 +1,2 @@
 from .setup import setup
+from .custom_formatter import CustomFormatter
