+++ rsmf/custom_formatter.py
@@ -1,6 +1,6 @@
 import matplotlib.pyplot as plt
-from .formatter import AbstractFormatter
-from .fontsize import default_fontsizes
+from .abstract_formatter import AbstractFormatter
+from .fontsizes import default_fontsizes
 
 
 class CustomFormatter(AbstractFormatter):
