+++ rsmf/quantumarticle_formatter.py
@@ -5,7 +5,7 @@ Support for the quantumarticle documentclass of Quantum journal.
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 
-from .revtexlike import RevtexLikeFormatter, RevtexLikeParser
+from .revtexlike_formatter import RevtexLikeFormatter, RevtexLikeParser
 
 
 class QuantumarticleFormatter(RevtexLikeFormatter):
